Move Test Only Metric from MergeTree Client to TestClient
