+++ packages/dds/merge-tree/package.json
@@ -109,14 +109,16 @@
     "baselineRange": ">=2.0.0-internal.2.0.0 <2.0.0-internal.3.0.0",
     "baselineVersion": "2.0.0-internal.2.1.1",
     "broken": {
-      "ClassDeclaration_Client": {"forwardCompat": false, "backCompat": false},
-      "InterfaceDeclaration_SegmentGroup": {"forwardCompat": false},
       "ClassDeclaration_Client": {
         "forwardCompat": false,
         "backCompat": false
       },
       "InterfaceDeclaration_SegmentGroup": {
         "forwardCompat": false
+      },
+      "InterfaceDeclaration_MergeTreeStats": {
+        "forwardCompat": false,
+        "backCompat": false
       }
     }
   }
