+++ packages/dds/merge-tree/src/client.ts
@@ -11,7 +11,7 @@ import { ISequencedDocumentMessage, MessageType } from "@fluidframework/protocol
 import { IFluidDataStoreRuntime, IChannelStorageService } from "@fluidframework/datastore-definitions";
 import { ISummaryTreeWithStats } from "@fluidframework/runtime-definitions";
 import type { IEventThisPlaceHolder, ITelemetryLogger } from "@fluidframework/common-definitions";
-import { assert, Trace, TypedEventEmitter, unreachableCase } from "@fluidframework/common-utils";
+import { assert, TypedEventEmitter, unreachableCase } from "@fluidframework/common-utils";
 import { LoggingError } from "@fluidframework/telemetry-utils";
 import { IIntegerRange } from "./base";
 import { RedBlackTree } from "./collections";
@@ -61,9 +61,7 @@ import {
     IMergeTreeDeltaOpArgs
 } from "./index";
 
-function elapsedMicroseconds(trace: Trace) {
-    return trace.trace().duration * 1000;
-}
+type IMergeTreeDeltaRemoteOpArgs = Omit<IMergeTreeDeltaOpArgs,"sequencedMessage"> & Required<Pick<IMergeTreeDeltaOpArgs, "sequencedMessage">>;
 
 /**
  * Emitted before this client's merge-tree normalizes its segments on reconnect, potentially
@@ -79,14 +77,6 @@ export interface IClientEvents {
 }
 
 export class Client extends TypedEventEmitter<IClientEvents> {
-    public measureOps = false;
-    public accumTime = 0;
-    public localTime = 0;
-    public localOps = 0;
-    public accumWindowTime = 0;
-    public accumWindow = 0;
-    public accumOps = 0;
-    public maxWindowTime = 0;
     public longClientId: string | undefined;
 
     private readonly _mergeTree: MergeTree;
@@ -250,23 +240,8 @@ export class Client extends TypedEventEmitter<IClientEvents> {
             segment);
 
         const opArgs = { op };
-        this.traceOp(
-            opArgs, 
-            () => this._mergeTree.insertAtReferencePosition(refPos, segment, opArgs)
-        );
-
+        this._mergeTree.insertAtReferencePosition(refPos, segment, opArgs);
         return op;
-    }
-
-    private traceOp(opArgs: IMergeTreeDeltaOpArgs, applyOp: () => void): void {
-        let traceStart: Trace | undefined;
-        if (this.measureOps) {
-            traceStart = Trace.start();
-        }
-
-        applyOp();
-
-        this.completeAndLogOp(opArgs, traceStart);
     }
 
     public walkSegments<TClientData>(
@@ -412,17 +387,15 @@ export class Client extends TypedEventEmitter<IClientEvents> {
         const op = opArgs.op;
         const clientArgs = this.getClientSequenceArgs(opArgs);
         const range = this.getValidOpRange(op, clientArgs);
-        this.traceOp(
-            opArgs,
-            () => this._mergeTree.markRangeRemoved(
-                range.start,
-                range.end,
-                clientArgs.referenceSequenceNumber,
-                clientArgs.clientId,
-                clientArgs.sequenceNumber,
-                false,
-                opArgs
-            )
+
+        this._mergeTree.markRangeRemoved(
+            range.start,
+            range.end,
+            clientArgs.referenceSequenceNumber,
+            clientArgs.clientId,
+            clientArgs.sequenceNumber,
+            false,
+            opArgs
         );
 
         return true;
@@ -443,18 +416,15 @@ export class Client extends TypedEventEmitter<IClientEvents> {
             return false;
         }
 
-        this.traceOp(
-            opArgs,
-            () => this._mergeTree.annotateRange(
-                range.start,
-                range.end,
-                op.props,
-                op.combiningOp,
-                clientArgs.referenceSequenceNumber,
-                clientArgs.clientId,
-                clientArgs.sequenceNumber,
-                opArgs
-            )
+        this._mergeTree.annotateRange(
+            range.start,
+            range.end,
+            op.props,
+            op.combiningOp,
+            clientArgs.referenceSequenceNumber,
+            clientArgs.clientId,
+            clientArgs.sequenceNumber,
+            opArgs
         );
 
         return true;
@@ -484,50 +454,19 @@ export class Client extends TypedEventEmitter<IClientEvents> {
             return false;
         }
 
-        this.traceOp(
-            opArgs,
-            () => this._mergeTree.insertSegments(
-                range.start,
-                segments!,
-                clientArgs.referenceSequenceNumber,
-                clientArgs.clientId,
-                clientArgs.sequenceNumber,
-                opArgs
-            )
-        );
 
+        this._mergeTree.insertSegments(
+            range.start,
+            segments,
+            clientArgs.referenceSequenceNumber,
+            clientArgs.clientId,
+            clientArgs.sequenceNumber,
+            opArgs
+        );
         return true;
     }
 
-    /**
-     *
-     * @param opArgs - The op args of the op to complete
-     * @param clientArgs - The client args for the op
-     * @param range - The range the op applied to
-     * @param clockStart - Optional. The clock start if timing data should be updated.
-     */
-    private completeAndLogOp(
-        opArgs: IMergeTreeDeltaOpArgs,
-        traceStart?: Trace) {
-        if (!opArgs.sequencedMessage) {
-            if (traceStart) {
-                this.localTime += elapsedMicroseconds(traceStart);
-                this.localOps++;
-            }
-        } else {
-            const { currentSeq, minSeq } = this.getCollabWindow();
-            const { minimumSequenceNumber, sequenceNumber } = opArgs.sequencedMessage;
-            assert(currentSeq < sequenceNumber,
-                0x030 /* "Incoming remote op sequence# <= local collabWindow's currentSequence#" */);
-            assert(minSeq <= minimumSequenceNumber,
-                0x031 /* "Incoming remote op minSequence# < local collabWindow's minSequence#" */);
-            if (traceStart) {
-                this.accumTime += elapsedMicroseconds(traceStart);
-                this.accumOps++;
-                this.accumWindow += (this.getCurrentSeq() - this.getCollabWindow().minSeq);
-            }
-        }
-    }
+
 
     /**
      * Returns a valid range for the op, or undefined
@@ -635,25 +574,15 @@ export class Client extends TypedEventEmitter<IClientEvents> {
         return this.getClientSequenceArgsForMessage(opArgs.sequencedMessage);
     }
 
-    private ackPendingSegment(opArgs: IMergeTreeDeltaOpArgs) {
-        const ackOp = (deltaOpArgs: IMergeTreeDeltaOpArgs) => {
-            let trace: Trace | undefined;
-            if (this.measureOps) {
-                trace = Trace.start();
-            }
-
+    private ackPendingSegment(opArgs: IMergeTreeDeltaRemoteOpArgs) {
+        const ackOp = (deltaOpArgs: IMergeTreeDeltaRemoteOpArgs) => {
             this._mergeTree.ackPendingSegment(deltaOpArgs);
             if (deltaOpArgs.op.type === MergeTreeDeltaType.ANNOTATE) {
                 if (deltaOpArgs.op.combiningOp && (deltaOpArgs.op.combiningOp.name === "consensus")) {
-                    this.updateConsensusProperty(deltaOpArgs.op, deltaOpArgs.sequencedMessage!);
+                    this.updateConsensusProperty(deltaOpArgs.op, deltaOpArgs.sequencedMessage);
                 }
             }
 
-            if (trace) {
-                this.accumTime += elapsedMicroseconds(trace);
-                this.accumOps++;
-                this.accumWindow += (this.getCurrentSeq() - this.getCollabWindow().minSeq);
-            }
         };
 
         if (opArgs.op.type === MergeTreeDeltaType.GROUP) {
@@ -788,10 +717,10 @@ export class Client extends TypedEventEmitter<IClientEvents> {
         return opList;
     }
 
-    private applyRemoteOp(opArgs: IMergeTreeDeltaOpArgs) {
+    private applyRemoteOp(opArgs: IMergeTreeDeltaRemoteOpArgs) {
         const op = opArgs.op;
         const msg = opArgs.sequencedMessage;
-        this.getOrAddShortClientId(msg!.clientId);
+        this.getOrAddShortClientId(msg.clientId);
         switch (op.type) {
             case MergeTreeDeltaType.INSERT:
                 this.applyInsertOp(opArgs);
@@ -849,7 +778,7 @@ export class Client extends TypedEventEmitter<IClientEvents> {
         this.getOrAddShortClientId(msg.clientId);
         // Apply if an operation message
         if (msg.type === MessageType.Operation) {
-            const opArgs: IMergeTreeDeltaOpArgs = {
+            const opArgs: IMergeTreeDeltaRemoteOpArgs = {
                 op: msg.contents as IMergeTreeOp,
                 sequencedMessage: msg,
             };
@@ -1027,18 +956,7 @@ export class Client extends TypedEventEmitter<IClientEvents> {
     }
 
     updateMinSeq(minSeq: number) {
-        let trace: Trace | undefined;
-        if (this.measureOps) {
-            trace = Trace.start();
-        }
         this._mergeTree.setMinSeq(minSeq);
-        if (trace) {
-            const elapsed = elapsedMicroseconds(trace);
-            this.accumWindowTime += elapsed;
-            if (elapsed > this.maxWindowTime) {
-                this.maxWindowTime = elapsed;
-            }
-        }
     }
 
     getContainingSegment<T extends ISegment>(pos: number, sequenceArgs?: Pick<ISequencedDocumentMessage, "referenceSequenceNumber" | "clientId">, localSeq?: number) {
